Allows analyzing n-gram data

It provides information on text files, such as number of lines and number of words. It displays bar plots showing n-gram frequencies.

Details

It provides a method that returns text file information. The text file information includes total number of lines, max, min and mean line length and file size. It provides a method that reads an input text file containing token frequencies. It displays the most occuring tokens.

Super class

wordpredictor::TextFileProcessor -> DataAnalyzer

Methods

Public methods

DataAnalyzer$new()

DataAnalyzer$plot_n_gram_stats()

DataAnalyzer$get_file_info()

DataAnalyzer$get_ngrams()

DataAnalyzer$clone()

Inherited methods

Method new()

It initializes the current object. It is used to set the file name and verbose options.

Usage

DataAnalyzer$new(fn = NULL, ve = 0)

Arguments

fn

The path to the input file.

ve

If progress information should be displayed.

Method plot_n_gram_stats()

It reads ngram token frequencies from an input text file. The ngram frequencies are then displayed in a bar plot. The type of plot is specified by the type option. 'top_features' displays the top n most occuring tokens along with their frequencies. 'coverage' displays the number of words along with their frequencies. The plot stats are returned by the function as a data frame.

Usage

DataAnalyzer$plot_n_gram_stats(opts)

Arguments

opts

The options for analyzing the data.

type. The type of plot to display. The options are: 'top_features', 'coverage'.

n. For 'top_features', it is the number of top most occuring tokens. For 'coverage' it is the first n frequencies.

save_to. The graphics devices to save the plot to. NULL implies plot is printed.

dir. The output directory where the plot will be saved.

Returns

A data frame containing the stats.

Method get_file_info()

Generates and returns information about text files.

Usage

DataAnalyzer$get_file_info(res)

Arguments

res

The name of a directory or a file name.

Returns

A data frame containing the text file statistics.

Method get_ngrams()

Returns the given number of ngrams and their frequencies. If the prefix parameter is not given, then the ngrams are randomly chosen. Otherise ngrams starting with the given regular expression are returned.

Usage

DataAnalyzer$get_ngrams(fn, c = NULL, pre = NULL)

Arguments

fn

The ngram file name.

c

The number of ngrams to return.

pre

The ngram prefix, given as a regular expression.

Method clone()

The objects of this class are cloneable with this method.

Usage

DataAnalyzer$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.