It performs extrinsic and intrinsic evaluation of a n-gram model

It provides methods for performing extrinsic and intrinsic evaluation. Intrinsic evaluation is based on calculation of Perplexity. Extrinsic evaluation involves determining the percentage of correct next word predictions.

Details

Before performing the intrinsic and extrinsic model evaluation, a validation file must be first generated. This can be done using the Generator class. Each line in the validation file is evaluated. For intrinsic evaluation Perplexity for the line is calculated. An overall summary of the Perplexity calculations is returned. It includes the min, max and mean Perplexity. For extrinsic evaluation, next word prediction is performed on each line. If the actual next word is one of the three predicted next words, then the prediction is considered to be accurate. The extrinsic evaluation returns the percentage of correct and incorrect predictions.

Super class

wordpredictor::TextFileProcessor -> ModelEvaluator

Methods

Public methods

ModelEvaluator$new()

ModelEvaluator$compare_performance()

ModelEvaluator$plot_stats()

ModelEvaluator$evaluate_performance()

ModelEvaluator$intrinsic_evaluation()

ModelEvaluator$extrinsic_evaluation()

ModelEvaluator$clone()

Inherited methods

Method new()

It initializes the current object. It is used to set the model file name and verbose options.

Usage

ModelEvaluator$new(mf = NULL, ve = 0)

Arguments

mf

The model file name.

ve

If progress information should be displayed.

Method compare_performance()

It compares the performance of the models in the given folder. The performance of the model is compared for the 4 metric which are time taken, memory used, Perplexity and accuracy. The performance comparision is displayed on plots. 4 plots are displayed. One for each performance metric. A fifth plot shows the variation of Perplexity with accuracy. All 5 plots are plotted on one page.

Usage

ModelEvaluator$compare_performance(opts)

Arguments

opts

The options for comparing model performance.

save_to. The graphics device to save the plot to. NULL implies plot is printed.

dir. The directory where the plot and stats will be saved.

mdir. The directory containing the model files.

Method plot_stats()

It plots the given stats on 5 plots. The plots are displayed on a single page. The 4 performance metrics which are time taken, memory, Perplexity and accuracy are plotted against the model name. Another plot compares Perplexity with accuracy for each model.

Usage

ModelEvaluator$plot_stats(data)

Arguments

data

The data to plot

Returns

The ggplot object is returned.

Method evaluate_performance()

It performs intrinsic and extrinsic evaluation for the given model. It also measures the memory usage and time taken. The performance metrics are displayed in 5 plots on one page. Performance statistics are saved to the model object.

Usage

ModelEvaluator$evaluate_performance(lc, fn)

Arguments

lc

The number of lines of text in the validation file to be used for the evaluation.

fn

The name of the validation file. If it does not exist, then the default file validation-clean.txt is checked in the models folder

Returns

The performance stats are returned.

Method intrinsic_evaluation()

Evaluates the model using intrinsic evaluation based on Perplexity. The given number of sentences are taken from the validation file. For each sentence, the Perplexity is calculated.

Usage

ModelEvaluator$intrinsic_evaluation(lc, fn)

Arguments

lc

The number of lines of text in the validation file to be used for the evaluation.

fn

The name of the validation file. If it does not exist, then the default file validation-clean.txt is checked in the models folder

Returns

The min, max and mean Perplexity score.

Method extrinsic_evaluation()

Evaluates the model using extrinsic evaluation based on Accuracy. The given number of sentences are taken from the validation file. For each sentence, the model is used to predict the next word. The accuracy stats are returned. A prediction is considered to be correct if one of the predicted words matches the actual word.

Usage

ModelEvaluator$extrinsic_evaluation(lc, fn)

Arguments

lc

The number of lines of text in the validation file to be used for the evaluation.

fn

The name of the validation file.

Returns

The number of correct and incorrect predictions.

Method clone()

The objects of this class are cloneable with this method.

Usage

ModelEvaluator$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.