It is used to generate transition probabilities for n-grams.

It provides a method for generating transition probabilities for the given n-gram size. It also provides a method for generating the combined transition probabilities data for n-gram sizes from 1 to the given size. The combined transition probabilities data can be used to implement back-off.

Details

It provides a method for generating n-gram transition probabilities. It reads ngram frequencies from an input text file that is generated by the TokenGenerator class. It parses each ngram into a prefix, a next word and the next word frequency. Maximum Likelihood count is used to generated the next word probabilities. The n-gram prefix is converted to a numeric hash. The next word is replaced with the position of the next word in the list of all words. The transition probabilities data is stored in a file. Another method is provided that combines the transition probabilities for n-grams of size 1 to the given size. The combined transition probabilities can be saved as a data frame. This file may be regarded as a n-gram model. The name of the file is model-n.RDS, where n is the n-gram number. By combining the transition probabilities of n-grams, back-off may be used to evaluate word probabilities or predict the next word.

Super class

wordpredictor::TextFileProcessor -> TPGenerator

Methods

Public methods

TPGenerator$new()

TPGenerator$generate_tp()

TPGenerator$generate_tp_for_n()

TPGenerator$clone()

Inherited methods

Method new()

It initializes the current obj. It is used to set the transition probabilities options and verbose option.

Usage

TPGenerator$new(opts = list(), ve = 0)

Arguments

opts

The options for generating the transition probabilities.

save_tp. If the data should be saved.

n. The ngram size.

dir. The directory containing the input and output files.

format. The format for the output. There are two options.

plain. The data is stored in plain text.

obj. The data is stored as a R obj.

ve

If progress information should be displayed.

Method generate_tp()

It combines the transition probabilities for n-grams less than or equal to the given n-gram size. It first generates the transition probabilities for each ngram size from 1 to the given size. The transition probabilities are then combined into a single data frame and saved to the output folder that is given as parameter to the current object. By combining the transition probabilities for all n-gram sizes less than n, back-off can be used to calculate next word probabilities or predict the next word.

Usage

TPGenerator$generate_tp()

Method generate_tp_for_n()

It reads ngram token frequencies from an input text file. It generates a data frame containing the prefix, next word and next word frequency. The data frame may be saved to a file as plain text or as a R obj. For n = 1, the list of words is saved.

Usage

TPGenerator$generate_tp_for_n(n)

Arguments

n

The n-gram size for which the tp data is generated.

Method clone()

The objects of this class are cloneable with this method.

Usage

TPGenerator$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.