Provides data cleaning functionality

It provides a memory efficient method for removing unneeded characters from text files. It is suitable for cleaning large text files.

Details

It provides a method for cleaning text files. It allows removing bad words, stop words, non dictionary words, extra space, punctuation and non-alphabet characters. Allows conversion to lower case. It supports large text files.

Super class

wordpredictor::TextFileProcessor -> DataCleaner

Methods

Public methods

DataCleaner$new()

DataCleaner$clean_file()

DataCleaner$clean_lines()

DataCleaner$clone()

Inherited methods

Method new()

It initializes the current object. It is used to set the file name and verbose options.

Usage

DataCleaner$new(fn = NULL, opts = list(), ve = 0)

Arguments

fn

The path to the file to clean.

opts

The options for data cleaning.

min_words. The minimum number of words per sentence.

line_count. The number of lines to read and clean at a time.

save_data. If the combined processed lines should be saved.

output_file. Name of the output file used to store the data.

sw_file. The stop words file path.

dict_file. The dictionary file path.

bad_file. The bad words file path.

to_lower. If the words should be converted to lower case.

remove_stop. If stop words should be removed.

remove_punct. If punctuation symbols should be removed.

remove_non_dict. If non dictionary words should be removed.

remove_non_alpha. -> If non alphabet symbols should be removed.

remove_extra_space. -> If leading, trailing and double spaces should be removed.

remove_bad. If bad words should be removed

ve

Indicates if progress information should be displayed.

Method clean_file()

It provides an efficient method for cleaning text files. It removes unneeded characters from the given text file with several options. It allows removing punctuations, bad words, stop words, non-alphabetical symbols and non-dictionary words. It reads a certain number of lines from the given text file. It removes unneeded characters from the lines and then saves the lines to an output text file. File cleaning progress is displayed if the verbose option was set in the class constructor. It is suitable for cleaning large text files.

Usage

DataCleaner$clean_file()

Method clean_lines()

It cleans the given lines of text using the options passed to the current object.

Usage

DataCleaner$clean_lines(lines)

Arguments

lines

The input sentences.

Returns

The cleaned lines of text.

Method clone()

The objects of this class are cloneable with this method.

Usage

DataCleaner$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.