It allows generating data samples from text files.

It provides a method for generating training, testing and validation data sets from a given input text file. It also provides a method for generating a sample file of given size or number of lines from an input text file.

Super class

wordpredictor::TextFileProcessor -> DataSampler

Methods

Public methods

DataSampler$new()

DataSampler$generate_sample()

DataSampler$generate_data()

DataSampler$clone()

Inherited methods

Method new()

It initializes the current object. It is used to set the verbose option.

Usage

DataSampler$new(ddir = "./data", mdir = "./models", ve = 0)

Arguments

ddir

The data directory.

mdir

The model directory.

ve

If progress information should be displayed.

Method generate_sample()

Generates a sample file of given size from the given input file. The file is optionally cleaned and saved to the data directory.

Usage

DataSampler$generate_sample(fn, ss, ic, ir, ofn, is)

Arguments

fn

The input file name. It is the short file name relative to the ddir. If not given, then the file name is auto generated from the type parameter.

ss

The number of lines or proportion of lines to sample.

ic

If the sample file should be cleaned.

ir

If the sample file contents should be randomized.

ofn

The output file name. It will be saved to the ddir.

is

If the sampled data should be saved to a file.

Method generate_data()

It generates training, testing and validation data sets from the given input file. It first reads the file given as a parameter to the current object. It partitions the data into training, testing and validation sets, according to the given parameters. The files are named train.txt, test.txt and va.txt. The files are saved to the given output folder.

Usage

DataSampler$generate_data(fn, dir, percs)

Arguments

fn

The input file name. It should be relative to the ddir.

dir

The name of the output folder.

percs

The size of the training, testing and validation sets.

Method clone()

The objects of this class are cloneable with this method.

Usage

DataSampler$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.