It is used to evaluate the accuracy and performance of the model.

It provides methods that perform extrinsic and intrinsic model evaluation. It also provides methods for determining the memory and time requirements for generating the model. It also provides a method for determining how much memory is used by the final model.

Details

It provides a method that performs intrinsic model evaluation based on Perplexity. It also provides a method that performs extrinsic model evalation based on accuracy. It provides a method for determining how much memory and time is needed to generate a model for different input data sizes. It provides a method for determining how much memory is needed by the final model.

Super class

wordpredictor::TextFileProcessor -> ModelPredictor

Methods

Public methods

ModelPredictor$new()

ModelPredictor$get_model()

ModelPredictor$calc_perplexity()

ModelPredictor$predict_word()

ModelPredictor$get_word_prob()

ModelPredictor$clone()

Inherited methods

Method new()

It initializes the current object. It is used to set the model file name and verbose options.

Usage

ModelPredictor$new(mf, ve = 0)

Arguments

mf

The model file name.

ve

If progress information should be displayed.

Method get_model()

Returns the Model class object.

Usage

ModelPredictor$get_model()

Returns

The Model class object is returned.

Method calc_perplexity()

The Perplexity for the given sentence is calculated. For each word, the probability of the word given the previous words is calculated. The probabilities are multiplied and then inverted. The nth root of the result is the perplexity, where n is the number of words in the sentence. If the stem_words tokenization option was specified, then the previous words are converted to their stem.

Usage

ModelPredictor$calc_perplexity(words)

Arguments

words

The list of words.

Returns

The perplexity of the given list of words.

Method predict_word()

Predicts the new word given a list of 1, 2 or 3 previous words. It checks the given n words in the transition probabilities data. If there is a match, the top 3 next words with highest probabilities are returned. If there is no match, then the last n-1 previous words are checked. This process is continued until the last word is checked. If there is no match, then empty result is returned. The given words may optionally be stemmed.

Usage

ModelPredictor$predict_word(words, count = 3, dc = NULL)

Arguments

words

A character vector of previous words or a single vector containing the previous word text.

count

The number of results to return.

dc

A DataCleaner object. If it is given, then the given words

Returns

The top 3 predicted words along with their probabilities.

Method get_word_prob()

Calculates the probability of the given word given the previous model-1 words, where model is the maximum ngram number. It looks up the probability of a word given n previous words. The previous n words are converted to numeric hash using digest2int function. The hash is looked up in a data frame of transition probabilities. The word is converted to a number by checking its position in a list of unique words. If the hash and the word position were found, then the probability of the previous word and hash is returned. If it was not found, then the hash of the n-1 previous words is taken and the processed is repeated. If the data was not found in the data frame, then the word probability is returned. This is known as backoff. If the word probability could not be found then the default probability is returned. The default probability is calculated as 1/(N+V), Where N = number of words in corpus and V is the number of dictionary words.

Usage

ModelPredictor$get_word_prob(word, pw)

Arguments

word

The word whoose probability is to be calculated.

pw

The previous words.

Returns

The probability of the word given the previous words.

Method clone()

The objects of this class are cloneable with this method.

Usage

ModelPredictor$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.